requirements.txt
# Core dependencies
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
markdown>=3.5.0,<4.0.0
PyYAML>=6.0.1,<7.0.0
rich>=13.7.0,<14.0.0
typer>=0.9.0,<1.0.0

# LangChain and LangGraph
langchain>=0.1.0,<0.3.0
langchain-community>=0.0.10,<0.3.0
langchain-core>=0.1.0,<0.3.0
langgraph>=0.0.20,<0.3.0

# LLM Providers
openai>=1.0.0,<2.0.0
anthropic>=0.8.0,<1.0.0
google-generativeai>=0.3.0,<1.0.0
requests>=2.31.0,<3.0.0

# Search Engines (使用更稳定的版本)
duckduckgo-search>=3.8.0,<3.9.0
google-search-results>=2.4.2,<3.0.0

# Web scraping and browser automation
beautifulsoup4>=4.12.0,<5.0.0
lxml>=4.9.0,<5.0.0
requests-html>=0.10.0,<1.0.0

# Browser-use for AI-powered browser automation
browser-use>=0.2.0,<1.0.0
playwright>=1.40.0,<2.0.0

# LangChain integrations for Browser-Use
langchain-openai>=0.1.0,<1.0.0
langchain-anthropic>=0.1.0,<1.0.0
langchain-google-genai>=1.0.0,<2.0.0

# Optional: Google Drive integration
# google-api-python-client>=2.100.0
# google-auth-httplib2>=0.1.1
# google-auth-oauthlib>=1.1.0

# Optional: Dropbox integration
# dropbox>=11.36.0

# Development and testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.5.0,<2.0.0

# Async support
aiohttp>=3.8.0,<4.0.0

# Docker support for sandboxing
docker>=6.0.0,<7.0.0 